clock + console + imports + poet

+++ src/shared/clock.hpp
@@ -2,7 +2,7 @@
 #include <shared/imports.hpp>
 
 namespace sharedtime {
-    typedef u64 timestamp;
+    typedef usize timestamp;
 
     namespace {
         inline void handle_overflow() {}
